+++ src/style-wrapper/style-wrapper.jsx
@@ -3,7 +3,9 @@ import CssBaseline from "@material-ui/core/CssBaseline";
 import React, { type Node } from "react";
 import { ThemeProvider } from "@material-ui/styles";
 import { createMuiTheme } from "@material-ui/core/styles";
-import { grailGrayLighter, grailPurpleDark, grailPurpleMed, grailRed, grailTan, white } from "@grail/lib";
+import {
+  grailGrayLighter, grailPurpleDark, grailPurpleMed, grailRed, grailTan, white,
+} from "@grail/lib";
 
 type Props = {
   children: Node,
@@ -24,6 +26,19 @@ const customTheme = createMuiTheme({
     error: {
       main: grailRed,
       contrastText: grailGrayLighter,
+    },
+  },
+  overrides: {
+    MuiButton: {
+      root: {
+        // Set the default color.
+        color: grailPurpleDark,
+      },
+    },
+  },
+  typography: {
+    button: {
+      fontWeight: 800,
     },
   },
 });